Introduced new statistics submodule, moved max, min, argmin and argmax to said module
