+++ heat/core/tests/test_relational.py
@@ -1,4 +1,3 @@
-import torch
 import unittest
 
 import heat as ht
@@ -164,159 +163,6 @@ class TestRelational(unittest.TestCase):
             ht.lt(self.a_tensor, self.errorneous_type)
         with self.assertRaises(TypeError):
             ht.lt('self.a_tensor', 's')
-
-    def test_max(self):
-        data = [
-            [1,   2,  3],
-            [4,   5,  6],
-            [7,   8,  9],
-            [10, 11, 12]
-        ]
-
-        ht_array = ht.array(data)
-        comparison = torch.tensor(data)
-
-        # check global max
-        maximum = ht.max(ht_array)
-
-        self.assertIsInstance(maximum, ht.Tensor)
-        self.assertEqual(maximum.shape, (1,))
-        self.assertEqual(maximum.lshape, (1,))
-        self.assertEqual(maximum.split, None)
-        self.assertEqual(maximum.dtype, ht.int64)
-        self.assertEqual(maximum._Tensor__array.dtype, torch.int64)
-        self.assertEqual(maximum, 12)
-
-        # maximum along first axis
-        maximum_vertical = ht.max(ht_array, axis=0)
-
-        self.assertIsInstance(maximum_vertical, ht.Tensor)
-        self.assertEqual(maximum_vertical.shape, (3,))
-        self.assertEqual(maximum_vertical.lshape, (1, 3,))
-        self.assertEqual(maximum_vertical.split, None)
-        self.assertEqual(maximum_vertical.dtype, ht.int64)
-        self.assertEqual(maximum_vertical._Tensor__array.dtype, torch.int64)
-        self.assertTrue((maximum_vertical._Tensor__array ==
-                         comparison.max(dim=0, keepdim=True)[0]).all())
-
-        # maximum along second axis
-        maximum_horizontal = ht.max(ht_array, axis=1)
-
-        self.assertIsInstance(maximum_horizontal, ht.Tensor)
-        self.assertEqual(maximum_horizontal.shape, (4,))
-        self.assertEqual(maximum_horizontal.lshape, (4, 1,))
-        self.assertEqual(maximum_horizontal.split, None)
-        self.assertEqual(maximum_horizontal.dtype, ht.int64)
-        self.assertEqual(maximum_horizontal._Tensor__array.dtype, torch.int64)
-        self.assertTrue((maximum_horizontal._Tensor__array == comparison.max(dim=1, keepdim=True)[0]).all())
-
-        # check max over all float elements of split 3d tensor, across split axis
-        random_volume = ht.random.randn(3, 3, 3, split=1)
-        maximum_volume = ht.max(random_volume, axis=1)
-
-        self.assertIsInstance(maximum_volume, ht.Tensor)
-        self.assertEqual(maximum_volume.shape, (3, 3))
-        self.assertEqual(maximum_volume.lshape, (3, 1, 3))
-        self.assertEqual(maximum_volume.dtype, ht.float32)
-        self.assertEqual(maximum_volume._Tensor__array.dtype, torch.float32)
-        self.assertEqual(maximum_volume.split, None)
-
-        # check max over all float elements of split 5d tensor, along split axis
-        random_5d = ht.random.randn(1, 2, 3, 4, 5, split=0)
-        maximum_5d = ht.max(random_5d, axis=1)
-
-        self.assertIsInstance(maximum_5d, ht.Tensor)
-        self.assertEqual(maximum_5d.shape, (1, 3, 4, 5))
-        self.assertLessEqual(maximum_5d.lshape[1], 2)
-        self.assertEqual(maximum_5d.dtype, ht.float32)
-        self.assertEqual(maximum_5d._Tensor__array.dtype, torch.float32)
-        self.assertEqual(maximum_5d.split, 0)
-
-        # check exceptions
-        with self.assertRaises(NotImplementedError):
-            ht_array.max(axis=(0, 1))
-        with self.assertRaises(TypeError):
-            ht_array.max(axis=1.1)
-        with self.assertRaises(TypeError):
-            ht_array.max(axis='y')
-        with self.assertRaises(ValueError):
-            ht.max(ht_array, axis=-4)
-
-    def test_min(self):
-        data = [
-            [1,   2,  3],
-            [4,   5,  6],
-            [7,   8,  9],
-            [10, 11, 12]
-        ]
-
-        ht_array = ht.array(data)
-        comparison = torch.tensor(data)
-
-        # check global max
-        minimum = ht.min(ht_array)
-
-        self.assertIsInstance(minimum, ht.Tensor)
-        self.assertEqual(minimum.shape, (1,))
-        self.assertEqual(minimum.lshape, (1,))
-        self.assertEqual(minimum.split, None)
-        self.assertEqual(minimum.dtype, ht.int64)
-        self.assertEqual(minimum._Tensor__array.dtype, torch.int64)
-        self.assertEqual(minimum, 12)
-
-        # maximum along first axis
-        minimum_vertical = ht.min(ht_array, axis=0)
-
-        self.assertIsInstance(minimum_vertical, ht.Tensor)
-        self.assertEqual(minimum_vertical.shape, (3,))
-        self.assertEqual(minimum_vertical.lshape, (1, 3,))
-        self.assertEqual(minimum_vertical.split, None)
-        self.assertEqual(minimum_vertical.dtype, ht.int64)
-        self.assertEqual(minimum_vertical._Tensor__array.dtype, torch.int64)
-        self.assertTrue((minimum_vertical._Tensor__array == comparison.min(dim=0, keepdim=True)[0]).all())
-
-        # maximum along second axis
-        minimum_horizontal = ht.min(ht_array, axis=1)
-
-        self.assertIsInstance(minimum_horizontal, ht.Tensor)
-        self.assertEqual(minimum_horizontal.shape, (4,))
-        self.assertEqual(minimum_horizontal.lshape, (4, 1,))
-        self.assertEqual(minimum_horizontal.split, None)
-        self.assertEqual(minimum_horizontal.dtype, ht.int64)
-        self.assertEqual(minimum_horizontal._Tensor__array.dtype, torch.int64)
-        self.assertTrue((minimum_horizontal._Tensor__array == comparison.min(dim=1, keepdim=True)[0]).all())
-
-        # check max over all float elements of split 3d tensor, across split axis
-        random_volume = ht.random.randn(3, 3, 3, split=1)
-        minimum_volume = ht.min(random_volume, axis=1)
-
-        self.assertIsInstance(minimum_volume, ht.Tensor)
-        self.assertEqual(minimum_volume.shape, (3, 3))
-        self.assertEqual(minimum_volume.lshape, (3, 1, 3))
-        self.assertEqual(minimum_volume.dtype, ht.float32)
-        self.assertEqual(minimum_volume._Tensor__array.dtype, torch.float32)
-        self.assertEqual(minimum_volume.split, None)
-
-        # check max over all float elements of split 5d tensor, along split axis
-        random_5d = ht.random.randn(1, 2, 3, 4, 5, split=0)
-        minimum_5d = ht.min(random_5d, axis=1)
-
-        self.assertIsInstance(minimum_5d, ht.Tensor)
-        self.assertEqual(minimum_5d.shape, (1, 3, 4, 5))
-        self.assertLessEqual(minimum_5d.lshape[1], 2)
-        self.assertEqual(minimum_5d.dtype, ht.float32)
-        self.assertEqual(minimum_5d._Tensor__array.dtype, torch.float32)
-        self.assertEqual(minimum_5d.split, 0)
-
-        # check exceptions
-        with self.assertRaises(NotImplementedError):
-            ht_array.min(axis=(0, 1))
-        with self.assertRaises(TypeError):
-            ht_array.min(axis=1.1)
-        with self.assertRaises(TypeError):
-            ht_array.min(axis='y')
-        with self.assertRaises(ValueError):
-            ht.min(ht_array, axis=-4)
 
     def test_ne(self):
         result = ht.uint8([
